Revert "Adding the pytests file and troubleshooting"
