--- pyproject.toml
@@ -1,46 +1,0 @@
-[tool.pytest.ini_options]
-minversion = "6.0"
-addopts = "-ra -q"
-testpaths = [
-    "tests",
-]
-
-[build-system]
-requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[tool.black]
-line-length = 88
-target-version = ['py39']
-
-[tool.isort]
-profile = "black"
-
-[project]
-name = "ccrypt"
-version = "0.1.0"
-description = "A package for Caesar cipher operations."
-authors = [
-    { name = "Your Name", email = "your.email@example.com" }
-]
-license = { text = "MIT" }
-readme = "README.md"
-requires-python = ">=3.9"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-]
-
-[project.urls]
-"Homepage" = "https://github.com/software-students-spring2024/3-python-package-exercise-ja-ia"
-"Repository" = "https://github.com/software-students-spring2024/3-python-package-exercise-ja-ia.git"
-"Bug Tracker" = "https://github.com/software-students-spring2024/3-python-package-exercise-ja-ia/issues"
-
-[project.scripts]
-ccrypt = "ccrypt.__main__:main"
-
-[tool.setuptools]
-packages = ["ccrypt"]
-package_dir = {""= "src"}
